requirements/development.txt
asgiref==3.7.2
black==24.3.0
boto3==1.34.68
botocore==1.34.68
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.3
django-allauth==0.61.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jmespath==1.0.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.4.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
